Accommodate Plugin Manager preview
The Plugin Manager assumes that the XML in the .sample is enough to
process the template. In the <flaw>'s element case we need to reference
data (cwename) that is higher up the XML hierarchy.

What we need is to provide the full XML hierarchy in the .sample, and
then have field_processor.rb determine whether it's been called from the
Importer (in which case a Flaw object can be created directly) or from
the Plugin Manager (in which case we have to create the object from the
right spot within the .sample XML).